refactor: re-organize setup files
