+++ common/services/fyle/org-settings.service.ts
@@ -1,5 +1,5 @@
-import { getRequestHeaders } from '../utils/api';
-import { FyleAccount } from './fyle-account';
+import { getRequestHeaders } from '../../utils/api';
+import { FyleAccount } from './fyle-account.service';
 
 export class OrgSettingsService {
   private readonly account: FyleAccount;
